stt almost moved to assistant ui
