+++ src/constants.js
@@ -5,7 +5,7 @@ export const STATUS = {
 	scheduled: 1,
 	unknown: 0,
 }
-export const TASK_TYPES = {
+export const TASK_CATEGORIES = {
 	text_generation: 0,
 	image_generation: 1,
 	speech_to_text: 2,
